refactor: move selfDestruct tests to SelfDestructSuite
